refactor(worker): extract GPU discovery into gpu.rs (sc-1636)
First worker production-seam split. Move the GPU-discovery cluster
(discover_gpu/discover_gpus, nvidia-smi querying + parsing, gpu_utilization,
visible-GPU env handling, cpu/fallback gpu, worker_capabilities*, and the
gpu/cpu worker-id helpers) out of lib.rs into gpu.rs.

Same pattern as the API split: `use super::*;` in the module (so the
shared DiscoveredGpu/contract types resolve unchanged), pub(crate) on the
fns reached from the run loop / supervisor / registration, glob re-import
at the crate root, and the worker tests.rs imports repointed to
super::gpu. DiscoveredGpu stays at the root (shared with the supervisor).
Behavior-preserving. lib.rs 4,854 -> 4,629 lines.
